Refactor constants to chatty.util package
