+++ src/main/java/chatty/ChattyChatBot.java
@@ -4,10 +4,27 @@ import chatty.task.Deadline;
 import chatty.task.Event;
 import chatty.task.Task;
 import chatty.task.ToDo;
+import chatty.util.Constants;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
+
+import static chatty.util.Constants.ADDED_TASK_CONFIRMATION;
+import static chatty.util.Constants.AT_STRING;
+import static chatty.util.Constants.BOT_NAME;
+import static chatty.util.Constants.BYE_STRING;
+import static chatty.util.Constants.BY_STRING;
+import static chatty.util.Constants.DEADLINE_STRING;
+import static chatty.util.Constants.DONE_STRING;
+import static chatty.util.Constants.DOT_CHARACTER;
+import static chatty.util.Constants.EVENT_STRING;
+import static chatty.util.Constants.LINE_BREAK;
+import static chatty.util.Constants.LIST_STRING;
+import static chatty.util.Constants.SPACE_SEPARATOR;
+import static chatty.util.Constants.TASK_SUMMARY_FIRST_HALF;
+import static chatty.util.Constants.TASK_SUMMARY_SECOND_HALF;
+import static chatty.util.Constants.TODO_STRING;
 
 public class ChattyChatBot {
 
@@ -22,32 +39,32 @@ public class ChattyChatBot {
         do {
             userInput = SCANNER.nextLine();
 
-            System.out.println(Constants.LINE_BREAK);
+            System.out.println(LINE_BREAK);
 
             // Solution below adapted from: https://stackoverflow
             // .com/questions/5067942/what-is-the-best-way-to-extract-the-first-word-from-a-string-in-java
-            String[] array = userInput.split(Constants.SPACE_SEPARATOR, 2);
+            String[] array = userInput.split(SPACE_SEPARATOR, 2);
             String action = array[0];
 
             switch (action) {
-            case Constants.LIST_STRING:
+            case LIST_STRING:
                 listAllTasks(tasks);
                 break;
-            case Constants.DONE_STRING:
+            case DONE_STRING:
                 try {
                     markTaskAsDone(tasks, array[1]);
                 } catch (ArrayIndexOutOfBoundsException e) {
                     System.out.println("Let me know which task you would like to mark as done?");
                 }
                 break;
-            case Constants.TODO_STRING:
+            case TODO_STRING:
                 try {
                     addToDoTask(tasks, array[1]);
                 } catch (ArrayIndexOutOfBoundsException e) {
                     System.out.println("Sure, let me know what ToDo task you would like to add!");
                 }
                 break;
-            case Constants.DEADLINE_STRING:
+            case DEADLINE_STRING:
                 try {
                     addDeadlineTask(tasks, array[1]);
                 } catch (ArrayIndexOutOfBoundsException e) {
@@ -56,7 +73,7 @@ public class ChattyChatBot {
                     System.out.println("Please specify the deadline of your task");
                 }
                 break;
-            case Constants.EVENT_STRING:
+            case EVENT_STRING:
                 try {
                     addEventTask(tasks, array[1]);
                 } catch (ArrayIndexOutOfBoundsException e) {
@@ -65,22 +82,22 @@ public class ChattyChatBot {
                     System.out.println("Please specify the time of your event");
                 }
                 break;
-            case Constants.BYE_STRING:
+            case BYE_STRING:
                 sendByeMessage();
                 break;
             default:
                 sendDefaultResponse();
             }
 
-            System.out.println(Constants.LINE_BREAK);
-        } while (!userInput.equals(Constants.BYE_STRING));
+            System.out.println(LINE_BREAK);
+        } while (!userInput.equals(BYE_STRING));
     }
 
     private static void listAllTasks(List<Task> tasks) {
         System.out.println("Here are the tasks in your list:");
         for (int i = 0; i < tasks.size(); i++) {
             Task task = tasks.get(i);
-            System.out.println((i + 1) + Constants.DOT_CHARACTER + Constants.SPACE_SEPARATOR + task.toString());
+            System.out.println((i + 1) + DOT_CHARACTER + SPACE_SEPARATOR + task.toString());
         }
     }
 
@@ -101,20 +118,20 @@ public class ChattyChatBot {
         ToDo newToDoTask = new ToDo(description);
         tasks.add(newToDoTask);
 
-        System.out.println(Constants.ADDED_TASK_CONFIRMATION);
+        System.out.println(ADDED_TASK_CONFIRMATION);
         System.out.println(newToDoTask);
-        System.out.println(Constants.TASK_SUMMARY_FIRST_HALF + tasks.size() + Constants.TASK_SUMMARY_SECOND_HALF);
+        System.out.println(TASK_SUMMARY_FIRST_HALF + tasks.size() + TASK_SUMMARY_SECOND_HALF);
     }
 
     private static void addDeadlineTask(List<Task> tasks, String inputStr) throws ChattyChatBotException {
         try {
-            String[] array = inputStr.split(Constants.BY_STRING);
+            String[] array = inputStr.split(BY_STRING);
             Deadline newDeadlineTask = new Deadline(array[0], array[1]);
             tasks.add(newDeadlineTask);
 
-            System.out.println(Constants.ADDED_TASK_CONFIRMATION);
+            System.out.println(ADDED_TASK_CONFIRMATION);
             System.out.println(newDeadlineTask);
-            System.out.println(Constants.TASK_SUMMARY_FIRST_HALF + tasks.size() + Constants.TASK_SUMMARY_SECOND_HALF);
+            System.out.println(TASK_SUMMARY_FIRST_HALF + tasks.size() + TASK_SUMMARY_SECOND_HALF);
         } catch (ArrayIndexOutOfBoundsException e) {
             throw new ChattyChatBotException();
         }
@@ -122,13 +139,13 @@ public class ChattyChatBot {
 
     private static void addEventTask(List<Task> tasks, String inputStr) throws ChattyChatBotException {
         try {
-            String[] array = inputStr.split(Constants.AT_STRING);
+            String[] array = inputStr.split(AT_STRING);
             Event newEventTask = new Event(array[0], array[1]);
             tasks.add(newEventTask);
 
-            System.out.println(Constants.ADDED_TASK_CONFIRMATION);
+            System.out.println(ADDED_TASK_CONFIRMATION);
             System.out.println(newEventTask);
-            System.out.println(Constants.TASK_SUMMARY_FIRST_HALF + tasks.size() + Constants.TASK_SUMMARY_SECOND_HALF);
+            System.out.println(TASK_SUMMARY_FIRST_HALF + tasks.size() + TASK_SUMMARY_SECOND_HALF);
         } catch (ArrayIndexOutOfBoundsException e) {
             throw new ChattyChatBotException();
         }
@@ -136,19 +153,19 @@ public class ChattyChatBot {
 
     private static void sendDefaultResponse() {
         System.out.println("Sorry, I can't help you with that yet");
-        System.out.println("I'm " + Constants.BOT_NAME);
+        System.out.println("I'm " + BOT_NAME);
         System.out.println("How may I help you?");
     }
 
     private static void sendWelcomeMessage() {
-        System.out.println(Constants.LINE_BREAK);
-        System.out.println("Hello from " + Constants.BOT_NAME);
+        System.out.println(LINE_BREAK);
+        System.out.println("Hello from " + BOT_NAME);
         System.out.println("Glad to be at your service!");
-        System.out.println(Constants.LINE_BREAK);
+        System.out.println(LINE_BREAK);
     }
 
     private static void sendByeMessage() {
-        System.out.println("Thanks for chatting with " + Constants.BOT_NAME);
+        System.out.println("Thanks for chatting with " + BOT_NAME);
         System.out.println("See you again soon!");
     }
 }
